Update OpenSC fork

+++ src/libopensc/errors.c
@@ -15,7 +15,7 @@
  *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
 #if HAVE_CONFIG_H
@@ -154,7 +154,7 @@ const char *sc_strerror(int error)
 
 	if (!error)
 		return no_errors;
-	error_index = error < 0 ? -error : error;
+	error_index = error < 0 ? -((long long int) error) : error;
 
 	if (error_index >= misc_base) {
 		errors = misc_errors;
